+++ packages/core-js/internals/function-uncurry-this.js
@@ -1,3 +1,4 @@
+var fails = require('../internals/fails');
 var NATIVE_BIND = require('../internals/function-bind-native');
 
 var FunctionPrototype = Function.prototype;
@@ -5,7 +6,12 @@ var bind = FunctionPrototype.bind;
 var call = FunctionPrototype.call;
 var uncurryThis = NATIVE_BIND && bind.bind(call, call);
 
-module.exports = NATIVE_BIND ? function (fn) {
+var UNCURRY_WITH_NATIVE_BIND = NATIVE_BIND && !fails(function () {
+  // Nashorn bug, https://github.com/zloirock/core-js/issues/1128
+  return uncurryThis(''.slice)('12', 1) !== '2';
+});
+
+module.exports = UNCURRY_WITH_NATIVE_BIND ? function (fn) {
   return fn && uncurryThis(fn);
 } : function (fn) {
   return fn && function () {